Move messaging to pkg
