+++ cmd/mqtt/main.go
@@ -15,12 +15,12 @@ import (
 	"github.com/go-redis/redis"
 	"github.com/mainflux/mainflux"
 	mflog "github.com/mainflux/mainflux/logger"
-	"github.com/mainflux/mainflux/messaging"
-	mqttpub "github.com/mainflux/mainflux/messaging/mqtt"
-	"github.com/mainflux/mainflux/messaging/nats"
 	"github.com/mainflux/mainflux/mqtt"
 	"github.com/mainflux/mainflux/mqtt/auth"
 	mqttredis "github.com/mainflux/mainflux/mqtt/redis"
+	"github.com/mainflux/mainflux/pkg/messaging"
+	mqttpub "github.com/mainflux/mainflux/pkg/messaging/mqtt"
+	"github.com/mainflux/mainflux/pkg/messaging/nats"
 	thingsapi "github.com/mainflux/mainflux/things/api/auth/grpc"
 	mp "github.com/mainflux/mproxy/pkg/mqtt"
 	"github.com/mainflux/mproxy/pkg/session"
