+++ http/api/metrics.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/go-kit/kit/metrics"
 	"github.com/mainflux/mainflux/http"
-	"github.com/mainflux/mainflux/messaging"
+	"github.com/mainflux/mainflux/pkg/messaging"
 )
 
 var _ http.Service = (*metricsMiddleware)(nil)
