+++ pkg/messaging/message.proto
