+++ pkg/messaging/nats/doc.go
