+++ twins/mocks/messages.go
@@ -5,7 +5,7 @@ package mocks
 
 import (
 	"github.com/mainflux/mainflux/errors"
-	"github.com/mainflux/mainflux/messaging"
+	"github.com/mainflux/mainflux/pkg/messaging"
 )
 
 var _ messaging.Publisher = (*mockBroker)(nil)
