+++ twins/mocks/service.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"time"
 
-	"github.com/mainflux/mainflux/messaging"
+	"github.com/mainflux/mainflux/pkg/messaging"
 	"github.com/mainflux/mainflux/twins"
 	"github.com/mainflux/mainflux/twins/uuid"
 	"github.com/mainflux/senml"
